feat(Scope:-Global,-Path:-src/**): Added the Cart Functionality, Modal Pop up, Animation on HeaderCart Button, State of the Items Added
Addl. : The Changes made are, a fully functional Add to CartButton is added which can add items to the Cart and let's say if you add a meal that's already in the Cart then the number of that particular meal will increase and not the whole meal will be added. Next, a modal of Cart is added which removes the whole cart on tapping the Remove Button if there was only one Amount of that element otherwise reduces the Amount Number of that Element. ALso, we can close the Cart Modal clicking on the Backdrop or on the Close Button ... Finally, while we add or remove Cart Items both from Modal or MealItems the HeaderCart Button makes a Bump Animation. That's the End of thisd Section ... @iamapurba2003

BREAKING CHANGE: Yes, REFERENCE ISSUES: N/A, Description: Chnages are made and new features are added -> [ Refer to Description ], Any mistakes, even a Small line Missing or wrong Data Type Operations or etc. could break the Functionality, Prone to Misbehave or in Worst Cases, might break the Whole App.