perf(Scope:-Global,-Path:-src/**): Internal Way of how React Works an Ways of Performance Improvments
Addl. : This Section describes how internal performance issues could happenn, in instances of State Changes and Entire Component Re - Evaluation, also hwo to resolve that with React.memo(), but also the fact that if value changes then the Re - Evaluation happens but contrary to that is that JavaScript functions are different every time you run it eve though its the same Source, hence if that's an Event Handler passed in a Component then the Entire Re - Evaluation Happens and to Save that Reat.useCallback() comes into Picture. But, if you fucntion in React.useCallback() has some other State Changing function out of the Scope then it won't work untill you add them as dependencies in the Dependency Array ... Moving Ahed, it's Worth noting that even though you can refer function to be the Same everytime you run it using React.useCallback(), if there are intensive operations going on in a Component and the whole Component Re - Evaluation is necessary then also you can restrict that particular part of Data to not Change and hence not Re - Evaluate that part using React.useMemo() Hook, basically Memoization technique is being taken into Picture .. Also, While making State Updates it's worth noticing that State Updates doesen't happens then and there, React Schedules State Update, it might be that internally there is some work going on which is more important that Updating State, i.e is why it Schedules State Update but yes it gurantees that the Order of State Updates remain Same as Specified, untill you are working in multiple Promise Blocks or ASynchronus ways. But it's worth noting that if there are multiple State Updates and React Schedules all of those that doesen't mean that React Re - Evaluates the Whole process for the Number of State Updates Scheduled, it batches them all into one State Updates with the Same Order a Soecified ... But to be in Safer side that you get latest presvious Snapshotof the Code you want in the State Update it's advisable to use the Function form in State updates. Finally, it's worth asking if State Updates leads to Re - Evaluation and Everything is re - created in Memory then how does State Variables remians same, it's because React.useState() renders the State Variable first time it's Created then it passes the Data to React to Store in it's internal Memory and React manages it there.

BREAKING CHANGE: No, REFERENCE ISSUES: N/A,  Description: These all are Performance Improvements it won't break the Code or Application untill Compute Overload or Server Error which is rare, but it might degrade the EPrformance of your Application if not Optimized using the Ways which is discussed in this Section briefly. Internal Customization is in hands of React but to Stabalize the Performance as much as possible is also a Work of a Good Developer.