perf(Scope:-Global,-Path:-src/**,-src/components/**): Added Error Boundaries Code in Class Based Component
Addl. : Added the COde for Error Boundaries which stops the frames and Callstack from collapsing, and Shows the Error Message instead which is defined by the User/Developer. It Comes in Handy whenever any Component Crashes and the Callstackas and Frames are COllapsed and throws Error. It improves the UI Experioence and Error Handling Better

BREAKING CHANGE: No. REFERENCE ISSUES: No, Description: Although, it isn't a Breaking Change still it plays important role in the UI Experience as you don't want the user to see the Error Screen in React rather a Custom Error Messege you provide, also Functional Components Doesen't Support Error Bounding, so for that thing to Handle you need a Class Based Component.